+++ packages/expo-dev-menu/android/src/debug/java/expo/modules/devmenu/DevMenuManager.kt
@@ -366,9 +366,8 @@ object DevMenuManager : DevMenuManagerInterface, LifecycleEventListener {
   }
 
   fun toggleFab() {
-    preferences?.showFab?.let {
-      preferences?.showFab = !it
-    }
+    val current = preferences?.showFab ?: return
+    preferences?.showFab = !current
   }
 
   override fun setDelegate(newDelegate: DevMenuDelegateInterface) {