+++ packages/expo-dev-menu/android/src/main/java/expo/modules/devmenu/compose/ui/DevMenuBottomSheet.kt
@@ -72,6 +72,7 @@ fun DevMenuBottomSheet(
       appInfo = appInfo,
       devToolsSettings = state.devToolsSettings,
       shouldShowOnboarding = shouldShowOnboarding.value,
+      showFab = state.showFab,
       onAction = wrappedOnAction
     )
   }